fixed matrix multiplicatino error when performin in-place assignment.

+++ actions/actions.h
@@ -4,10 +4,10 @@
 
 typedef struct dct_matrix_t dct_matrix_t;
 
-void dct_compress(const char* path, int d, int p);
+void dct_compress(const char* path, int d, float p);
 void dct_expand(const char* path);
 
 void dct_compute_transform(dct_matrix_t *result, int d);
-void dct_compute_quantize(dct_matrix_t *result, int d, int p);
+void dct_compute_quantize(dct_matrix_t *result, int d, float p);
 
 #endif//DCT_MODULES_ACTIONS_ACTIONS_H_
